fabfile: Chef query uses `chef_environment`
The Chef query should use `chef_environment` instead of `environment`:
http://wiki.opscode.com/display/chef/Environments#Environments-SearchingwithinEnvironments.
The `environment` search should have not worked to begin with, but
it has only recently realized that and stopped working. This commit
will fix deploys that are failing with the "No hosts found. Please
specify (single) host string for connection" error.