pylintrc-test-overrides: Disable annoying messages
This pull request will change the error messages "Too many public
methods" and "Catching Exception" from being generated.

The reason for disabling these errors is that simply inheriting from
unittest2.TestCase will trigger the "Too many public methods".  Since
we are exclusively using unittest2 in favor of dingus this will cause
each of our test cases to generate an error message.

The reason for disabling the second message, "Catching Exception" is that
in the EvO unit tests we catch a general error then inspect various aspects
about that error to make sure it was raised from the right context and with
the correct message.