Create proper package for fabric files.
Representing the fabric control file as a Python module is problematic
if anyone downstream needs to modify it for any reason.  By switching
the fabfile to a package it can be extended downstream while preserving
the ability to push out changes from base_services.  As a bonus, this
also makes it possible to introduce testing of our fabric infrastructure
should it become that complex.