Makefile: Rework testing targets.
The existing testing targets were overly complex and relied on target
specific variables to implement generic targets.  This set of commits
normalizes the testing and coverage targets so that they use the pattern
matching facilities provided by Make instead.  This makes for a more
readable makefile though it still does rely on some Makefile-voodoo.

The rework of the testing targets includes one subtle change.  Instead
of using `setup.py -q test` to run the tests, I switched to using the
nose runner explicitly.  The primary reason for this change is that the
former relied on a command alias in setup.cfg and it included the
additional overhead of pulling in `test_requires` packages from setup.py
on every test run.